backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
openai
python-multipart
langchain_community
pinecone

# ML and Data Science libraries
transformers
torch
scikit-learn
pandas
numpy
sentence-transformers

# Additional utilities
aiofiles
python-multipart
pydantic